:recycle: move coreNamespace to its own file
