+++ packages/core/src/operation/core-namespace.ts
@@ -1,0 +1,1 @@
+export const coreNamespace = '@cuillere/core'
