+++ packages/core/src/operation/generator.ts
@@ -1,4 +1,5 @@
-import { Operation, coreNamespace } from './operation'
+import { coreNamespace } from './core-namespace'
+import { Operation } from './operation'
 
 const generatorOperation: Operation = Object.freeze({ kind: `${coreNamespace}/generator` })
