+++ packages/core/src/operation/index.ts
@@ -1,5 +1,6 @@
 export * from './call'
 export * from './callable'
+export * from './core-namespace'
 export * from './defer'
 export * from './execute'
 export * from './fork'
