+++ packages/core/src/operation/terminal.ts
@@ -1,5 +1,6 @@
+import { coreNamespace } from './core-namespace'
 import { Effect } from '../effect'
-import { coreNamespace, isOfKind } from './operation'
+import { isOfKind } from './operation'
 import { WrapperOperation } from './wrapper'
 
 const kind = `${coreNamespace}/terminal`
